Update XrefAssociationAdaptor.pm
Slight change to the SQL to remove the un-required link to the gene_tree_root table at the root_id is in the gene_tree_node table and not referenced else where.

Also I have removed the e.db_name from the group by as you are already filtering by the external_db for a single record already. This just adds extra computation that is not required.